created List.js and NewTodo.js
