+++ src/App.css
@@ -1,16 +1,24 @@
-.App {
-  text-align: center;
+body {
+  font: 14px 'Helvetica Neue', Helvetica, Arial, sans-serif;
+  font-weight: normal;
+  line-height: normal;
+  line-height: 1.4em;
+  color: #4d4d4d;
+  font-weight: 300;
 }
 
-.App-header {
-  background-color: #282c34;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  color: white;
+.input-field {
+  font-size: 18px;
+  padding: 10px 4px;
 }
 
-.App-list {
+.todos {
   list-style: none;
+  font-size: 18px;
+  margin-left: 0;
+  padding-left: 0;
+}
+
+.todo--completed {
+  text-decoration: line-through;
 }
